Move most flatbuffers util to ort_flatbuffers
