+++ onnxruntime/core/graph/model.cc
@@ -3,9 +3,9 @@
 
 #include <memory>
 #include "core/common/logging/logging.h"
-#include "core/flatbuffers/ort.fbs.h"
+#include "core/flatbuffers/schema/ort.fbs.h"
+#include "core/flatbuffers/flatbuffers_utils.h"
 #include "core/framework/tensorprotoutils.h"
-#include "core/graph/graph_flatbuffers_utils.h"
 #include "core/graph/model.h"
 #include "core/graph/model_load_utils.h"
